+++ Echo/EchoPlugin.cs
@@ -67,11 +67,17 @@ namespace Echo
             AlphaPlugin.CommandManager?.Register(new EchoLocationListCommand());
             AlphaPlugin.CommandManager?.Register(new EchoLocationRemoveCommand());
             AlphaPlugin.CommandManager?.Register(new EchoFollowCommand());
+            AlphaPlugin.CommandManager?.Register(new EchoFollowRelativeCommand());
             AlphaPlugin.CommandManager?.Register(new EchoUnfollowCommand());
             AlphaPlugin.CommandManager?.Register(new EchoSyncRotationCommand());
             AlphaPlugin.CommandManager?.Register(new EchoUnsyncRotationCommand());
+            AlphaPlugin.CommandManager?.Register(new EchoRotateCommand());
+            AlphaPlugin.CommandManager?.Register(new EchoRotateLockCommand());
+            AlphaPlugin.CommandManager?.Register(new EchoRotateUnlockCommand());
+            AlphaPlugin.CommandManager?.Register(new EchoLookAtCommand());
             AlphaPlugin.CommandManager?.Register(new EchoCopyNameCommand());
             AlphaPlugin.CommandManager?.Register(new EchoCopyNameStyleCommand());
+            AlphaPlugin.CommandManager?.Register(new EchoSetNameCommand());
             AlphaPlugin.CommandManager?.Register(new EchoGetNameCommand());
             AlphaPlugin.CommandManager?.Register(new EchoRevertNameCommand());
             AlphaPlugin.CommandManager?.Register(new EchoCopyOutfitCommand());